back/requirements.txt
fastapi[all]>=0.75
uvicorn[standard]>=0.15
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
bcrypt>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5  # PostgreSQL用
pydantic-settings
pytest==8.3.4